requirement.txt
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
Markdown==3.3.4
pycparser==2.20
PyMySQL==1.0.2
pytz==2021.1
sqlparse==0.4.1
